Create game class
